ci(pipeline): Upload debug artifacts
The scraping spider is saving debug HTML files to help determine the
correct CSS selectors for job detail pages.

This commit updates the GitHub Actions workflow to correctly upload the
`src/pipeline/debug/` directory as part of the `pipeline-artifacts`.
This will allow for the analysis of the saved HTML files.